setup database migrations
