+++ requirements.txt
@@ -1,13 +1,16 @@
+alembic==1.11.1
 blinker==1.6.2
 click==8.1.3
 exceptiongroup==1.1.1
 Flask==2.3.2
 Flask-Cors==3.0.10
+Flask-Migrate==4.0.4
 Flask-SQLAlchemy==3.0.3
 greenlet==2.0.2
 iniconfig==2.0.0
 itsdangerous==2.1.2
 Jinja2==3.1.2
+Mako==1.2.4
 MarkupSafe==2.1.3
 packaging==23.1
 pluggy==1.0.0
