+++ rockr/models/__init__.py
@@ -1,1 +1,13 @@
 from .user import User
+from .match_profile import (
+    MatchProfile,
+    UserMatch,
+    MusicalInterest,
+    UserMusicalInterest,
+    Instrument,
+    UserInstrument,
+    Goal,
+    UserGoal
+)
+from .band import Band, UserBand
+from.message import Message, UserMessageGroup
