+++ rockr/models/user.py
@@ -1,7 +1,6 @@
 import rockr.auth0.auth0_api_wrapper as auth0
 from rockr import db
 from .mixins import SerializerMixin
-from .match_profile import Goal
 
 
 class User(SerializerMixin, db.Model):
@@ -20,31 +19,3 @@ class User(SerializerMixin, db.Model):
         # this acts like a property and can be accessed like user.user_role
         api_wrapper = auth0.Auth0ApiWrapper()
         return api_wrapper.get_user_role(self.id)
-
-
-class UserMatch(SerializerMixin, db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('public.user.id'),
-                        nullable=False)  # user receiving the match
-    match_id = db.Column(db.Integer, db.ForeignKey('public.user.id'),
-                         nullable=False)  # user that is the match
-    accepted = db.Column(db.Boolean, default=False)
-    seen = db.Column(db.Boolean, default=False)
-
-
-class UserMusicalInterest(SerializerMixin, db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('public.user.id'), nullable=False)
-    interest_id = db.Column(db.Integer, db.ForeignKey('public.musicalinterest.id'), nullable=False)
-
-
-class UserInstrument(SerializerMixin, db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('public.user.id'), nullable=False)
-    instrument_id = db.Column(db.Integer, db.ForeignKey('public.instrument.id'), nullable=False)
-
-
-class UserGoal(SerializerMixin, db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-    goal_id = db.Column(db.Integer, db.ForeignKey(Goal.id), nullable=False)
