--- rockr/rockr_schema.sql
@@ -1,228 +1,0 @@
---
--- PostgreSQL database dump
---
-
--- Dumped from database version 15.3
--- Dumped by pg_dump version 15.3
-
-SET statement_timeout = 0;
-SET lock_timeout = 0;
-SET idle_in_transaction_session_timeout = 0;
-SET client_encoding = 'UTF8';
-SET standard_conforming_strings = on;
-SELECT pg_catalog.set_config('search_path', '', false);
-SET check_function_bodies = false;
-SET xmloption = content;
-SET client_min_messages = warning;
-SET row_security = off;
-
-SET default_tablespace = '';
-
-SET default_table_access_method = heap;
-
---
--- Name: bands; Type: TABLE; Schema: public; Owner: postgres
---
-
-CREATE TABLE public.bands (
-    pkid integer NOT NULL,
-    band_name text,
-    genre text,
-    members integer
-);
-
-
-ALTER TABLE public.bands OWNER TO postgres;
-
---
--- Name: bands_pkid_seq; Type: SEQUENCE; Schema: public; Owner: postgres
---
-
-CREATE SEQUENCE public.bands_pkid_seq
-    AS integer
-    START WITH 1
-    INCREMENT BY 1
-    NO MINVALUE
-    NO MAXVALUE
-    CACHE 1;
-
-
-ALTER TABLE public.bands_pkid_seq OWNER TO postgres;
-
---
--- Name: bands_pkid_seq; Type: SEQUENCE OWNED BY; Schema: public; Owner: postgres
---
-
-ALTER SEQUENCE public.bands_pkid_seq OWNED BY public.bands.pkid;
-
-
---
--- Name: roles; Type: TABLE; Schema: public; Owner: postgres
---
-
-CREATE TABLE public.roles (
-    pkid integer NOT NULL,
-    role_name text
-);
-
-
-ALTER TABLE public.roles OWNER TO postgres;
-
---
--- Name: roles_pkid_seq; Type: SEQUENCE; Schema: public; Owner: postgres
---
-
-CREATE SEQUENCE public.roles_pkid_seq
-    AS integer
-    START WITH 1
-    INCREMENT BY 1
-    NO MINVALUE
-    NO MAXVALUE
-    CACHE 1;
-
-
-ALTER TABLE public.roles_pkid_seq OWNER TO postgres;
-
---
--- Name: roles_pkid_seq; Type: SEQUENCE OWNED BY; Schema: public; Owner: postgres
---
-
-ALTER SEQUENCE public.roles_pkid_seq OWNED BY public.roles.pkid;
-
-
---
--- Name: users; Type: TABLE; Schema: public; Owner: postgres
---
-
-CREATE TABLE public.users (
-    pkid integer NOT NULL,
-    username text,
-    first_name text,
-    last_name text
-);
-
-
-ALTER TABLE public.users OWNER TO postgres;
-
---
--- Name: users_pkid_seq; Type: SEQUENCE; Schema: public; Owner: postgres
---
-
-CREATE SEQUENCE public.users_pkid_seq
-    AS integer
-    START WITH 1
-    INCREMENT BY 1
-    NO MINVALUE
-    NO MAXVALUE
-    CACHE 1;
-
-
-ALTER TABLE public.users_pkid_seq OWNER TO postgres;
-
---
--- Name: users_pkid_seq; Type: SEQUENCE OWNED BY; Schema: public; Owner: postgres
---
-
-ALTER SEQUENCE public.users_pkid_seq OWNED BY public.users.pkid;
-
-
---
--- Name: bands pkid; Type: DEFAULT; Schema: public; Owner: postgres
---
-
-ALTER TABLE ONLY public.bands ALTER COLUMN pkid SET DEFAULT nextval('public.bands_pkid_seq'::regclass);
-
-
---
--- Name: roles pkid; Type: DEFAULT; Schema: public; Owner: postgres
---
-
-ALTER TABLE ONLY public.roles ALTER COLUMN pkid SET DEFAULT nextval('public.roles_pkid_seq'::regclass);
-
-
---
--- Name: users pkid; Type: DEFAULT; Schema: public; Owner: postgres
---
-
-ALTER TABLE ONLY public.users ALTER COLUMN pkid SET DEFAULT nextval('public.users_pkid_seq'::regclass);
-
-
---
--- Data for Name: bands; Type: TABLE DATA; Schema: public; Owner: postgres
---
-
-COPY public.bands (pkid, band_name, genre, members) FROM stdin;
-1	Def Leppard	80s Metal	5
-2	Sum 41	Pop Punk	5
-\.
-
-
---
--- Data for Name: roles; Type: TABLE DATA; Schema: public; Owner: postgres
---
-
-COPY public.roles (pkid, role_name) FROM stdin;
-1	basic_user
-2	admin
-3	band_lead
-\.
-
-
---
--- Data for Name: users; Type: TABLE DATA; Schema: public; Owner: postgres
---
-
-COPY public.users (pkid, username, first_name, last_name) FROM stdin;
-1	shults21	larry	schultheis
-\.
-
-
---
--- Name: bands_pkid_seq; Type: SEQUENCE SET; Schema: public; Owner: postgres
---
-
-SELECT pg_catalog.setval('public.bands_pkid_seq', 2, true);
-
-
---
--- Name: roles_pkid_seq; Type: SEQUENCE SET; Schema: public; Owner: postgres
---
-
-SELECT pg_catalog.setval('public.roles_pkid_seq', 3, true);
-
-
---
--- Name: users_pkid_seq; Type: SEQUENCE SET; Schema: public; Owner: postgres
---
-
-SELECT pg_catalog.setval('public.users_pkid_seq', 1, true);
-
-
---
--- Name: bands bands_pkey; Type: CONSTRAINT; Schema: public; Owner: postgres
---
-
-ALTER TABLE ONLY public.bands
-    ADD CONSTRAINT bands_pkey PRIMARY KEY (pkid);
-
-
---
--- Name: roles roles_pkey; Type: CONSTRAINT; Schema: public; Owner: postgres
---
-
-ALTER TABLE ONLY public.roles
-    ADD CONSTRAINT roles_pkey PRIMARY KEY (pkid);
-
-
---
--- Name: users users_pkey; Type: CONSTRAINT; Schema: public; Owner: postgres
---
-
-ALTER TABLE ONLY public.users
-    ADD CONSTRAINT users_pkey PRIMARY KEY (pkid);
-
-
---
--- PostgreSQL database dump complete
---
-
